backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
pymysql==1.0.2
cryptography==3.4.7
